Parser refactor to add builder and player name fix
