+++ ts/src/parser.types.ts
@@ -1,0 +1,6 @@
+export type GameRawData = {
+    gameCount: number,
+    totalKills: number,
+    players: Set<string>,
+    kills: Record<string, number>
+}
